pyproject.toml
[build-system]
requires = ["flit_core >= 3.4"]
build-backend = "flit_core.buildapi"

[project]
name = "metrabs_pytorch"
version = "0.1.0"
authors = [{name = ""}]
description = "A Python package for human motion tracking"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  # "torch",
  # "torchvision",
  "hydra-core",
  "ultralytics",
  "av",
  "posepile@git+https://github.com/Vittorio-Caggiano/PosePile.git",
  "cameralib@git+https://github.com/isarandi/cameralib.git",
  "mediapy",
  "einops",
  "tqdm",
  "Pillow<10",
  "matplotlib",
]

[project.optional-dependencies]
viz = [
  "poseviz@git+https://github.com/isarandi/poseviz.git",
]
gradio = [
  "gradio>=4.0",
]

[tool.ruff.lint]
ignore = ["C901", "E501", "E741", "F402", "F823" ]
select = ["C", "E", "F", "I", "W"]
